+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 ## [Unreleased]
 
+### Added
+
+- Add support for Spot VMs for Azure Node Pools.
+
 ## [1.32.0] - 2021-07-16
 
 ### Changed